# use generics instead of dynamic dispatch internally in `tar::Builder`