**TL;DR** Using generics instead of dynamic dispatch allows for specialization and a performance win on modern Linux systems.

| bench                          | old time       | new time        | delta   |
|:-------------------------------|:---------------|:----------------|--------:|
| cache=yes size=583M files=23k  | 1.9 s ± 0.1 s  | 1.7 s ± 0.1 s   | -10.08% |
| cache=no  size=583M files=23k  | 7.3 s ± 0.2 s  | 6.7 s ± 0.3 s   |  -8.95% |
| cache=yes size=284M files=1k   | 500 ms ±  4 ms | 362 ms ± 28 ms  | -38.12% |
| cache=no  size=284M files=1k   | 918 ms ± 25 ms | 889 ms ± 29 ms  |  -3.26% |

(look [here](https://gist.github.com/GabrielDertoni/ac6832046e6eabec43d4e45da5a3b027) for full benchmark details)

This PR is specifically targeted to allow triggering the specialization of the `Write` implementations when using `tar::Builder`. In particular, the [stdlib has a specialization on Linux for `std::io::copy`](https://github.com/rust-lang/rust/blob/7bb98889534d597911263d647424557b414770ec/library/std/src/sys/pal/unix/kernel_copy.rs#L74) that allows it to use the [`copy_file_range` syscall](https://man7.org/linux/man-pages/man2/copy_file_range.2.html). Unfortunately the code as-is is not capable of performing such specialization (or potentially others) since it relies on dynamic dispatch internally. This PR changes the underlying implementation to uses generics instead of dynamic dispatch. This understandably comes at the potential cost of increased code size but I'd argue this choice is already exposed to users that should be willing to pay the code size cost when using `tar::Builder<T>` directly instead of `tar::Builder<dyn Write>`. Users looking for reduced code footprint can explicitly opt in through using `tar::Builder<dyn Write>` instead.