@cgwalters 2025-01-15
> In particular, the [stdlib has a specialization on Linux for std::io::copy](https://github.com/rust-lang/rust/blob/7bb98889534d597911263d647424557b414770ec/library/std/src/sys/pal/unix/kernel_copy.rs#L74) that allows it to use the [copy_file_range syscall](https://man7.org/linux/man-pages/man2/copy_file_range.2.html).

And that would happen here when your tar output is to a file, and you're providing a `File` instance as input e.g. `append()`? OK. (Or a pipe/socket for either of those)

OK, though I do kind of question how often this occurs, because *most* use cases for tar end up being compressed (i.e. the output side is not a fd).

Can you at briefly explain your higher level use case and what led you to try this optimization?