@GabrielDertoni 2025-01-17
Fair point on the fact that this is mostly used for compression purposes. However, this specialization also applies to writers  like `BufWriter<T>` which can have their larger internal buffers used directly, avoiding one unnecessary copy.

I stumbled on this when someone told me they were trying to archive a python environment and taring was taking a significant portion of the time. I then looked through the implementation and realized many syscalls were being issued that could be significantly reduced. Although this optimization may dramatically reduce the number of syscalls in some cases it seldom results in as big of a time reduction, but it is not negligible.

This can be quite useful whey you'd like to write to an archive temporarily, even if it is going to be compressed afterwards, or you want to pipe that result into another program to handle the compression. I guess what surprised me is that since `tar::Builder` is a generic type my performance expectation is that it will monorphize and benefit where it can from that so I found it quite odd that internally it uses dynamic dispatch.