@cgwalters 2025-01-17
There are certainly use cases for uncompressed tar; over in OCI land we were discussing that uncompressed tar should used when the data embedded is already compressed - AI models can be like this. This helps avoid a problem in OCI because it checksums compressed data, and that can make it hard to ensure reproducible caching.

Anyways though...the change sounds reasonable but, this crate is *widely* used and I personally try to be pretty conservative with what we merge. This type of change has the potential to cause regressions (probably in code size mostly - but certainly not exclusively).

Can you please provide a bit more benchmarking in a targeted use case?

> I stumbled on this when someone told me they were trying to archive a python environment and taring was taking a significant portion of the time. I then looked through the implementation and realized many syscalls were being issued that could be significantly reduced. Although this optimization may dramatically reduce the number of syscalls in some cases it seldom results in as big time reduction, but it is not negligible.

A bit more data? Are we talking 1%? 10%? As far as syscall traffic, of course others have been looking at io-uring based work (see https://github.com/alexcrichton/tar-rs/pull/355 ), but that's a much deeper change.