@GabrielDertoni 2025-01-19
Hi, I understand you being conservative and I appreciate it! Unfortunately I am not able to provide benchmarks more specific than I have already provided in the original description, but the perf benefits are in the order showed on the previous benchmarks (up to 38%, but 10% mostly). I am certainly looking at io-uring based alternatives myself but I agree that's a much larger change and may not fit in this crate. This PR though I thought would benefit upstream for a sensible optimization without too many downsides. And I'd be interested to know what other downsides you think this would cause downstream and if there is some specific example you might be worried about. I'd also be happy to consider feature-flagging this if you think that'd be more appropriate.