@NobodyXu 2025-01-19
> I am certainly looking at io-uring based alternatives myself but I agree that's a much larger change and may not fit in this crate.

IIRC there are desires to extract and create a sans-IO core tar crate